Небольшой рефакторинг модуля git - execute-git вынесен в отдельный файл и перенесен в папку libs - git разбит на несколько модулей: git-file-content.js, git-file-tree.js, git-history.js - API не изменен: всё экспортится через index.js в папке git
