+++ package.json
@@ -16,6 +16,7 @@
     "hbs": "^4.0.1"
   },
   "devDependencies": {
+    "jest": "^23.6.0",
     "ndb": "^1.0.26"
   }
 }
