+++ utils/git/git-file-content.js
@@ -1,0 +1,6 @@
+const root = require('path').resolve('.')
+const executeGit = require(root + '/libs/execute-git')
+
+module.exports = function gitFileContent(hash) {
+  return executeGit('git', ['show', hash]);
+}
