+++ utils/git/git-file-tree.js
@@ -1,0 +1,23 @@
+const root = require('path').resolve('.')
+const executeGit = require(root + '/libs/execute-git')
+
+function parseFileTreeItem(line) {
+  const [info, path] = line.split('\t');
+  const [, type, hash] = info.split(' ');
+
+  return { type, hash, path };
+}
+
+function gitFileTree(hash, path) {
+  const params = ['ls-tree', hash];
+  path && params.push(path);
+
+  return executeGit('git', params).then(data => {
+    return data
+      .split('\n')
+      .filter(Boolean)
+      .map(parseFileTreeItem);
+  });
+}
+
+module.exports = gitFileTree
