+++ utils/git/git-history.js
@@ -1,0 +1,34 @@
+const root = require('path').resolve('.')
+const executeGit = require(root + '/libs/execute-git')
+
+function parseHistoryItem(line) {
+  const [hash, author, timestamp, msg] = line.split('\t');
+
+  return {
+    hash,
+    author,
+    timestamp,
+    msg
+  };
+}
+
+function gitHistory(page = 1, size = 10) {
+  const offset = (page - 1) * size;
+
+  return executeGit('git', [
+    'log',
+    '--pretty=format:%H%x09%an%x09%ad%x09%s',
+    '--date=iso',
+    '--skip',
+    offset,
+    '-n',
+    size
+  ]).then(data => {
+    return data
+      .split('\n')
+      .filter(Boolean)
+      .map(parseHistoryItem);
+  });
+}
+
+module.exports = gitHistory
