+++ utils/git/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  gitHistory: require('./git-history'),
+  gitFileTree: require('./git-file-tree'),
+  gitFileContent: require('./git-file-content')
+}
